Add files via upload
A challenge in the banking industry is to retain their customers. Every bank wants to hold their customers for sustaining their business so does our client - The ABC Multinational bank. Predicting the churn (of customer) in a multi-national banking sector is a real-life problem that needs to be solved in a more accurate and automated way. There are several factors that results in churning of a customer. Banking sectors apply numerous techniques for analysing and predicting if a customer is going to churn or not.